requirements/base.txt
fastapi==0.95.2
psycopg2==2.9.6
requests==2.31.0
sqlalchemy==2.0.15
uvicorn==0.22.0
python-dotenv==1.0.0